app/requirements.txt
flask==2.0.1
opencv-python==4.5.3.56
numpy==1.21.2
tensorflow==2.9.1
mediapipe==0.8.9.1
flask-socketio==5.1.1
python-telegram-bot==13.7 